Refactor first time zoom logic into super class
