+++ SampleApp/DemoRouteViewController.swift
@@ -15,8 +15,6 @@ class DemoRouteViewController: SampleMapViewController, MapSingleTapGestureDeleg
   let routeListSegueId = "routeListSegueId"
   var currentRouteResult: OTRRoutingResult?
   var lastRoutingPoint: OTRGeoPoint?
-  var firstTimeZoomToCurrentLocation = true
-  var sceneDidLoad = false
 
   private var routingLocale = Locale.current {
     didSet {
@@ -112,14 +110,6 @@ class DemoRouteViewController: SampleMapViewController, MapSingleTapGestureDeleg
     }
   }
 
-  func shouldZoomToCurrentLocation() {
-    if !sceneDidLoad { return }
-    if lastSetPoint == nil { return }
-    _ = resetCameraOnCurrentLocation()
-    firstTimeZoomToCurrentLocation = false
-  }
-
-
   //MARK:- Single Tap Gesture Recognizer Delegate
   func mapController(_ controller: MZMapViewController, recognizer: UIGestureRecognizer, shouldRecognizeSingleTapGesture location: CGPoint) -> Bool {
     return true
@@ -128,13 +118,5 @@ class DemoRouteViewController: SampleMapViewController, MapSingleTapGestureDeleg
   func mapController(_ controller: MZMapViewController, recognizer: UIGestureRecognizer, didRecognizeSingleTapGesture location: CGPoint) {
     let point = tgViewController.screenPosition(toLngLat: location)
     requestRoute(toPoint: OTRGeoPoint(coordinate: point))
-  }
-
-  //MARK:- Location Delegate Overrides
-  override func locationDidUpdate(_ location: CLLocation) {
-    super.locationDidUpdate(location)
-    if (firstTimeZoomToCurrentLocation) {
-      shouldZoomToCurrentLocation()
-    }
   }
 }
