+++ SampleApp/SampleMapViewController.swift
@@ -7,10 +7,28 @@
 //
 
 import Foundation
+import CoreLocation
 
 class SampleMapViewController : MZMapViewController {
 
+  var firstTimeZoomToCurrentLocation = true
+  var sceneDidLoad = false
+
   override var prefersStatusBarHidden: Bool {
     return false
+  }
+  func shouldZoomToCurrentLocation() {
+    if !sceneDidLoad { return }
+    if lastSetPoint == nil { return }
+    _ = resetCameraOnCurrentLocation()
+    firstTimeZoomToCurrentLocation = false
+  }
+
+  //MARK:- Location Delegate Overrides
+  override func locationDidUpdate(_ location: CLLocation) {
+    super.locationDidUpdate(location)
+    if (firstTimeZoomToCurrentLocation) {
+      shouldZoomToCurrentLocation()
+    }
   }
 }
